bracket naming refactoring
