+++ sway-fmt-v2/src/utils.rs
@@ -1,3 +1,3 @@
+pub mod bracket;
 pub mod indent_style;
 pub mod newline_style;
-pub mod punctuation;
