+++ sway-fmt-v2/src/utils/bracket.rs
